api_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pyjwt
requests
kafka-python
pydantic
python-dotenv
docker
